requirements.txt
aiohttp==3.12.13
discord.py==2.5.2
